requirements.txt
Bootstrap_Flask
Flask
flask_sqlalchemy
Flask_WTF
Requests
WTForms
werkzeug
python-dotenv